Update tarball names in hashes
When the `-linux` suffix was added these were not updated
resulting in same noise. We should probably come up with a `die`
implementation that works within this on the off chance something
goes wrong, but that can wait.

Refactor to use `pushd` and `popd` to save on a few `cd`s and avoid
some subshells.

Closes: #3
Closes: #4
Closes: #6
Closes: #7